[Snap-1286] Store side changes proposed in this pull request (#151)
Changes for snap-1286

+++ gemfirexd/client/src/main/java/com/pivotal/gemfirexd/internal/client/am/DatabaseMetaData.java
@@ -2375,10 +2375,10 @@ public abstract class DatabaseMetaData implements java.sql.DatabaseMetaData {
         try
         {
 			if (metaDataInfoIsCached_) {
-				return ((Integer) metaDataInfoCache_[infoCallIndex]).intValue() != 0;
+				return ((int) metaDataInfoCache_[infoCallIndex]) != 0;
 			}
 			metaDataInfoCall();
-			return ((Integer) metaDataInfoCache_[infoCallIndex]).intValue() != 0;
+			return ((int) metaDataInfoCache_[infoCallIndex]) != 0;
         }
         catch ( SqlException se )
         {
